+++ src/types.h
@@ -23,11 +23,11 @@
 #include <inttypes.h>
 
 
-typedef unsigned short int  dimm_t;
-typedef ssize_t             pos_t;
-typedef int_least8_t        bool_t;
-typedef int_least8_t        byte_t;
-typedef int_least32_t       char_t;
+typedef size_t         dimm_t;
+typedef ssize_t        pos_t;
+typedef int_least8_t   bool_t;
+typedef int_least8_t   byte_t;
+typedef int_least32_t  char_t;
 
 
 #endif
